requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-binance>=1.0.19
requests>=2.31.0
flask>=3.0.0

